chore(*): 예외 처리 클래스 세분화
